Split ProcessTest and move to unit-tests/shared
build.sbt was also adjusted to accomodate adding envVars to JVM.
